+++ client.go
@@ -6,7 +6,6 @@ import (
 	"errors"
 	"fmt"
 	"io"
-	"io/ioutil"
 	"log"
 	"math/rand"
 	"net/http"
@@ -65,7 +64,7 @@ func (r *lockedRand) Float64() float64 {
 // other hand, the source returned from rand.NewSource is not safe for
 // concurrent use, so we need to couple its use with a sync.Mutex.
 var rng = &lockedRand{
-	//#nosec G404 -- We are fine using transparent, non-secure value here.
+	// #nosec G404 -- We are fine using transparent, non-secure value here.
 	r: rand.New(rand.NewSource(time.Now().UnixNano())),
 }
 
@@ -79,7 +78,7 @@ type usageError struct {
 
 // Logger is an instance of log.Logger that is use to provide debug information about running Sentry Client
 // can be enabled by either using Logger.SetOutput directly or with Debug client option.
-var Logger = log.New(ioutil.Discard, "[Sentry] ", log.LstdFlags)
+var Logger = log.New(io.Discard, "[Sentry] ", log.LstdFlags)
 
 // EventProcessor is a function that processes an event.
 // Event processors are used to change an event before it is sent to Sentry.
@@ -178,7 +177,8 @@ type ClientOptions struct {
 	Dist string
 	// The environment to be sent with events.
 	Environment string
-	// Maximum number of breadcrumbs.
+	// Maximum number of breadcrumbs
+	// when MaxBreadcrumbs is negative then ignore breadcrumbs.
 	MaxBreadcrumbs int
 	// Maximum number of spans.
 	//
@@ -352,6 +352,10 @@ func (client *Client) setupIntegrations() {
 		integration.SetupOnce(client)
 		Logger.Printf("Integration installed: %s\n", integration.Name())
 	}
+
+	sort.Slice(client.integrations, func(i, j int) bool {
+		return client.integrations[i].Name() < client.integrations[j].Name()
+	})
 }
 
 // AddEventProcessor adds an event processor to the client. It must not be
@@ -402,7 +406,7 @@ func (client *Client) Recover(err interface{}, hint *EventHint, scope EventModif
 	// use the Context for communicating deadline nor cancelation. All it does
 	// is store the Context in the EventHint and there nil means the Context is
 	// not available.
-	//nolint: staticcheck
+	// nolint: staticcheck
 	return client.RecoverWithContext(nil, err, hint, scope)
 }
 
@@ -596,22 +600,22 @@ func (client *Client) prepareEvent(event *Event, hint *EventHint, scope EventMod
 	}
 
 	if event.ServerName == "" {
-		if client.Options().ServerName != "" {
-			event.ServerName = client.Options().ServerName
-		} else {
+		event.ServerName = client.Options().ServerName
+
+		if event.ServerName == "" {
 			event.ServerName = hostname
 		}
 	}
 
-	if event.Release == "" && client.Options().Release != "" {
+	if event.Release == "" {
 		event.Release = client.Options().Release
 	}
 
-	if event.Dist == "" && client.Options().Dist != "" {
+	if event.Dist == "" {
 		event.Dist = client.Options().Dist
 	}
 
-	if event.Environment == "" && client.Options().Environment != "" {
+	if event.Environment == "" {
 		event.Environment = client.Options().Environment
 	}
 
@@ -655,11 +659,10 @@ func (client *Client) prepareEvent(event *Event, hint *EventHint, scope EventMod
 }
 
 func (client Client) listIntegrations() []string {
-	integrations := make([]string, 0, len(client.integrations))
-	for _, integration := range client.integrations {
-		integrations = append(integrations, integration.Name())
+	integrations := make([]string, len(client.integrations))
+	for i, integration := range client.integrations {
+		integrations[i] = integration.Name()
 	}
-	sort.Strings(integrations)
 	return integrations
 }